Skip godep generation for commits containing incorrect godep changes
This commit adds an exception list for commits containing bad godep changes
(eg client-go depending on apiserver) and skips godep generation for these
commits.

However, we make sure to reset godeps for each commit (including the bad commit).
The correct godep generation will take place once it encounters a good commit.

This is a temporary fix and should probably be reverted because it is not always
true that there will be a commit having the correct godep changes after a bad one
_immediately_. If there isn't a good commit after a bad one immediately, the commits
between them will have bad Godeps.json.